+++ src/main/java/com/flash21/accounting/contract/service/ContractServiceImpl.java
@@ -57,7 +57,7 @@ public class ContractServiceImpl implements ContractService {
     }
 
     @Override
-    public ContractResponseDto getContractById(Integer contractId) {
+    public ContractResponseDto getContractById(Long contractId) {
         Contract contract = contractRepository.findById(contractId)
                 .orElseThrow(() -> new AccountingException(ContractErrorCode.NOT_FOUND));
         return toResponseDto(contract);
@@ -65,7 +65,7 @@ public class ContractServiceImpl implements ContractService {
 
     @Override
     @Transactional
-    public ContractResponseDto updateContract(Integer contractId, ContractRequestDto requestDto) {
+    public ContractResponseDto updateContract(Long contractId, ContractRequestDto requestDto) {
         Contract contract = contractRepository.findById(contractId)
                 .orElseThrow(() -> new AccountingException(ContractErrorCode.NOT_FOUND));
 
@@ -87,7 +87,7 @@ public class ContractServiceImpl implements ContractService {
     }
 
     @Override
-    public void deleteContract(Integer contractId) {
+    public void deleteContract(Long contractId) {
         if (!contractRepository.existsById(contractId)) {
             throw new AccountingException(ContractErrorCode.NOT_FOUND);
         }
